moved bitset manipulation so separate file
