+++ include/l3ster/util/Common.hpp
@@ -39,54 +39,20 @@ constexpr bool exactlyOneOf(T... args) requires(std::convertible_to< T, bool >an
     return (static_cast< size_t >(static_cast< bool >(args)) + ...) == 1u;
 }
 
-template < size_t n_bits >
-constexpr size_t bitsetNUllongs()
+// Workaround: GCC Bug 97930
+// Needed as template parameter because libstdc++ std::pair is not structural (private base class)
+// TODO: rely on std::pair once fixed upstream
+template < typename T1, typename T2 >
+struct Pair
 {
-    constexpr auto ull_bits = sizeof(unsigned long long) * 8u;
-    return n_bits / ull_bits + static_cast< bool >(n_bits % ull_bits);
-}
+    using first_type  = T1;
+    using second_type = T2;
 
-template < std::size_t N >
-auto serializeBitset(const std::bitset< N >& bits)
-{
-    using ull                    = unsigned long long;
-    constexpr auto ull_bits      = sizeof(ull) * 8u;
-    constexpr auto required_ulls = bitsetNUllongs< N >();
-    const auto     mask          = std::bitset< N >{std::numeric_limits< ull >::max()}; // 0xffffffffffffffff
+    constexpr Pair() requires(std::is_default_constructible_v< T1 >and std::is_default_constructible_v< T2 >) = default;
+    constexpr Pair(T1 t1, T2 t2) : first{std::move(t1)}, second{std::move(t2)} {}
 
-    std::array< ull, required_ulls > retval;
-    std::generate_n(
-        rbegin(retval), retval.size(), [&, i = 0ul]() mutable { return (bits >> i++ * ull_bits & mask).to_ullong(); });
-    return retval;
-}
-
-template < std::size_t N >
-auto deserializeBitset(const std::array< unsigned long long, N >& data)
-{
-    using ull               = unsigned long long;
-    constexpr auto ull_bits = sizeof(ull) * 8u;
-    using ret_t             = std::bitset< N * ull_bits >;
-    ret_t retval;
-    for (auto chunk64 : data)
-    {
-        retval <<= ull_bits;
-        retval |= ret_t{chunk64};
-    }
-    return retval;
-}
-
-template < std::size_t N_out, std::size_t N_in >
-auto trimBitset(const std::bitset< N_in >& in) requires(N_out <= N_in)
-{
-    if constexpr (N_in <= sizeof(unsigned long long) * 8u)
-        return std::bitset< N_out >{in.to_ullong()};
-    else
-    {
-        std::bitset< N_out > retval;
-        for (std::size_t i = 0u; i < N_out; ++i)
-            retval[i] = in[i];
-        return retval;
-    }
-}
+    T1 first;
+    T2 second;
+};
 } // namespace lstr
 #endif // L3STER_UTIL_COMMON_HPP
